ReferenceDescriptor, SubmodelDescriptor: interface and Default implementation
